+++ CHANGELOG.rst
@@ -3,6 +3,7 @@
 
 **Features**:
 
+* Add usage of certificate chains in opc ua stack #401, by @huebl
 * Add new thread and communication concept #339, by @huebl 
 * Add configuration file to opc ua client application #301, by @huebl
 * Replace slot timer callback function by timeoutfunction #285, by @huebl